Split off hinter functionality into a separate file
